WIP Fix problem with items not being removed from the scene when paused. Updated duration logic to use double. Make sure we only process one schedule change event at once. Prevent startNext on paused Layouts.

+++ Logic/InterruptState.cs
@@ -25,11 +25,11 @@ namespace XiboClient.Logic
 {
     class InterruptState
     {
-        public int SecondsInterrutedThisHour = 0;
+        public double SecondsInterrutedThisHour = 0;
         public int TargetHourlyInterruption = 0;
         public DateTime LastInterruption;
         public DateTime LastPlaytimeUpdate;
         public DateTime LastInterruptScheduleChange;
-        public Dictionary<int, int> InterruptTracking;
+        public Dictionary<int, double> InterruptTracking;
     }
 }
